Naprawiono błąd związany z dodawaniem i usuwaniem zdjęć produktów

+++ backend/services/fileService.js
@@ -1,16 +1,27 @@
 import { ProductImage } from '../config/dbConfig.js';
 import fs from 'fs';
 import path from 'path';
+import { fileURLToPath } from 'url';
 
-const UPLOADS_DIR = 'public/images';
-
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+const UPLOADS_DIR = path.join(__dirname, '..', 'public', 'images');
 
 export async function uploadImages(files) {
-    const uploadedFilesInfo = files.map(file => {
-        return `/images/${file.filename}`;
-    });
+    try {
+        const uploadedFilesInfo = files.map(file => {
+            const filePath = path.join(UPLOADS_DIR, file.filename);
+            const tempPath = file.path;
 
-    return uploadedFilesInfo;
+            fs.renameSync(tempPath, filePath);
+
+            return `/images/${file.filename}`;
+        });
+        return uploadedFilesInfo;
+    } catch (err) {
+        console.error('Błąd podczas przesyłania zdjęć:', err);
+        throw err;
+    }
 };
 
 export async function deleteImage(imageId) {
@@ -21,8 +32,13 @@ export async function deleteImage(imageId) {
         }
 
         const filePath = path.join(__dirname, '..', 'public', image.imagePath);
-        
-        fs.unlinkSync(filePath);
+
+        if (fs.existsSync(filePath)) {
+            fs.unlinkSync(filePath);
+        } else {
+            console.warn('Plik nie istnieje:', filePath);
+        }
+
         await image.destroy();
         return { message: 'Obrazek został usunięty.' };
     } catch (err) {
